Rename sailthru_client to client
